# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agenvas API
  version: 0.1.0
  description: |
    Agenvas 的权威 HTTP 合约。覆盖当前已实现的身份、项目、画布、Agent、Task 与私有素材切片；未实现的 MVP 能力继续按开发清单加入。
    部署者设置 AGENVAS_RECOVERY_MODE=true 后，/api/v1/projects、/api/v1/settings 与 /api/v1/media-templates 下的非 GET/HEAD/OPTIONS 请求统一返回 HTTP 503 ProblemDetail，code=RECOVERY_MODE_READ_ONLY；身份登录与只读核对仍可用。恢复模式同时关闭后台调度，不能把 503 当作业务版本冲突重试。
    服务端为每次 HTTP 请求生成 X-Request-Id 响应头；ProblemDetail.traceId 与它相同。客户端同名请求头不会被信任或回显。
    所有 API 接受 Accept-Language 语言协商（en、zh、ru、ja，地区变体按基础语言匹配；不支持、缺失或无效时回退 zh，q=0 不选用）。响应携带 Content-Language 与 Vary: Accept-Language，包括认证和 CSRF 错误。Problem 的 title、detail、fieldErrors.message 使用协商语言；code、状态、字段路径、ID 和用户内容不随语言变化。语言偏好仅由浏览器保存，无账号配置 API 或数据库迁移。
servers:
  - url: /
paths:
...
  "/api/v1/projects/{projectId}/canvas-items/{itemId}/media-draft/replace-inputs":
    "parameters":
      - {"$ref":"#/components/parameters/AcceptLanguage"}
      - {"name":"projectId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}
      - {"name":"itemId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}
    "post":
      "operationId": "replaceMediaDraftInputs"
      "tags":
        - "MediaTemplates"
      "responses":
        "200":
          "description": "Success"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/MediaDraft"
        "400":
          "$ref": "#/components/responses/ValidationError"
        "401":
          "$ref": "#/components/responses/Unauthenticated"
        "403":
          "$ref": "#/components/responses/Forbidden"
        "404":
          "$ref": "#/components/responses/NotFound"
        "409":
          "$ref": "#/components/responses/Conflict"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/SaveMediaDraftRequest"
      "description": "Atomically removes current media-input connection lines and saves the complete replacement draft with expectedVersion CAS. Preserves the previous draft and lines on conflict or validation failure. Does not generate media."
...
  /api/v1/projects/{projectId}/canvas-items/{canvasItemId}/media-draft:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - {$ref: "#/components/parameters/ProjectId"}
      - {$ref: "#/components/parameters/CanvasItemId"}
    get:
      operationId: getMediaDraft
      summary: 读取卡片独立的图片或视频工作草稿
      tags: [Canvas]
      responses:
        "200":
          description: 当前工作草稿
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaDraft"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
    put:
      operationId: saveMediaDraft
      summary: 按卡片草稿版本号保存媒体工作草稿
      tags: [Canvas]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SaveMediaDraftRequest"}
      responses:
        "200":
          description: 已保存的草稿
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaDraft"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/projects/{projectId}/canvas-items/{canvasItemId}/media-draft/restore-version-inputs:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - {$ref: "#/components/parameters/ProjectId"}
      - {$ref: "#/components/parameters/CanvasItemId"}
    post:
      operationId: restoreMediaDraftVersionInputs
      summary: 用历史版本的冻结输入完整替换卡片草稿
      tags: [Canvas]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RestoreMediaDraftVersionInputsRequest"}
      responses:
        "200":
          description: 已恢复的媒体草稿；图片只保留手工来源
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaDraft"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/projects/{projectId}/canvas-items/{canvasItemId}/media-draft/media-inputs/{versionId}/remove:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - {$ref: "#/components/parameters/ProjectId"}
      - {$ref: "#/components/parameters/CanvasItemId"}
      - name: versionId
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      operationId: removeMediaDraftMediaInput
      summary: 原子移除精确图片输入、全部来源连线及其结构化标签
      tags: [Canvas]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RemoveMediaDraftMediaInputRequest"}
      responses:
        "200":
          description: 已移除图片输入的媒体草稿
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaDraft"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  schemas:
...
    RunningHubScalar:
      oneOf:
        - {type: string, maxLength: 20000}
        - {type: number}
        - {type: boolean}
...
    SaveMediaDraftRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, prompt, parameters, videoInputMode, mediaInputs, mentions]
      properties:
        expectedVersion: {type: integer, format: int64, minimum: 0}
        prompt: {type: string, maxLength: 20000}
        parameters: {$ref: "#/components/schemas/ImageGenerationParameters"}
        durationSeconds: {type: [integer, "null"], minimum: 1, maximum: 60, description: "RunningHub 的具名时长来源支持 1–60 整数秒；其他适配器仍按各自较小范围校验。未声明时长字段的 RunningHub 视频可以为 null。"}
        capabilityId: {type: [string, "null"], format: uuid}
        styleId: {type: [string, "null"], format: uuid, description: Image/video only. Backend appends the selected preset after rendering media mentions; prompt remains user text. Disabled selections fail submission explicitly.}
        videoInputMode:
          oneOf:
            - {$ref: "#/components/schemas/VideoInputMode"}
            - {type: "null"}
        mediaInputs:
          type: array
          maxItems: 8
          items: {$ref: "#/components/schemas/SaveMediaInput"}
        mentions:
          type: array
          items: {$ref: "#/components/schemas/PromptMediaMention"}
    RestoreMediaDraftVersionInputsRequest:
      type: object
      additionalProperties: false
      required: [versionId, expectedVersion]
      properties:
        versionId: {type: string, format: uuid}
        expectedVersion: {type: integer, format: int64, minimum: 0}
    RemoveMediaDraftMediaInputRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: {type: integer, format: int64, minimum: 0}
...
    MediaDraft:
      type: object
      additionalProperties: false
      required: [projectId, canvasItemId, prompt, parameters, durationSeconds, capabilityId, styleId,
        videoInputMode, mediaInputs, mentions, displayMode, version, createdAt, updatedAt]
      properties:
        projectId: {type: string, format: uuid}
        canvasItemId: {type: string, format: uuid}
        prompt: {type: string, maxLength: 20000}
        parameters: {$ref: "#/components/schemas/ImageGenerationParameters"}
        durationSeconds: {type: [integer, "null"], minimum: 1, maximum: 60, description: "RunningHub 的具名时长来源支持 1–60 整数秒；其他适配器仍按各自较小范围校验。未声明时长字段的 RunningHub 视频可以为 null。"}
        capabilityId: {type: [string, "null"], format: uuid}
        styleId: {type: [string, "null"], format: uuid, description: Image/video only. Backend appends the selected preset after rendering media mentions; prompt remains user text. Disabled selections fail submission explicitly.}
        videoInputMode:
          oneOf:
            - {$ref: "#/components/schemas/VideoInputMode"}
            - {type: "null"}
        mediaInputs:
          type: array
          items: {$ref: "#/components/schemas/MediaInput"}
        mentions:
          type: array
          items: {$ref: "#/components/schemas/PromptMediaMention"}
        displayMode: {type: string, enum: [DRAFT, RESULT]}
        version: {type: integer, format: int64, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    ImageGenerationParameters:
      type: object
      additionalProperties: false
      description: 媒体草稿的原子生成参数；图片使用图片字段，视频使用 aspectRatio 和 AutoDL 专用 videoResolution，音频使用 speaker 和 speechRate。缺省字段由服务端按兼容默认值补齐并在 Task 中冻结。
      properties:
        dynamicValues:
          type: object
          maxProperties: 78
          additionalProperties: {$ref: "#/components/schemas/RunningHubScalar"}
          description: RunningHub named inputs or ComfyUI declared scalar inputs and reference_0 through reference_13 slots. Media values are exact version UUIDs matched against deduplicated mediaInputs; the same version may supply multiple named slots. ComfyUI drafts without the dynamicValues property retain legacy positional assignment; an explicit empty object means all slots are unassigned. Arbitrary graph targets and URLs are rejected.
        speaker: {type: string, pattern: "^[A-Za-z0-9_-]{0,120}$"}
        speechRate: {type: integer, minimum: -50, maximum: 100}
        loudnessRate: {type: integer, minimum: -50, maximum: 100}
        pitchRate: {type: integer, minimum: -12, maximum: 12}
        aspectRatio: {type: string, enum: [AUTO, "1:1", "2:3", "3:2", "9:16", "16:9", "3:4", "4:3", "21:9"]}
        videoResolution: {type: string, pattern: '^[1-9][0-9]{2,3}p$', description: "AutoDL only; a published resolution tier. Omitted uses the capability default. Frozen into accepted tasks."}
        resolution: {type: string, enum: [1K, 2K, 4K]}
        quality: {type: string, enum: [low, medium, high]}
        transparentBackground: {type: boolean}
        generationCount: {type: integer, enum: [1, 2, 4]}
    VideoInputMode:
      type: string
      enum: [TEXT, START_END, GENERAL_REFERENCE]
    MediaInputRole:
      type: string
      enum: [REFERENCE, START_FRAME, END_FRAME, AUDIO_REFERENCE, VIDEO_REFERENCE]
    SaveMediaInput:
      type: object
      additionalProperties: false
      required: [versionId, role, color]
      properties:
        versionId: {type: string, format: uuid}
        role: {$ref: "#/components/schemas/MediaInputRole"}
        color: {type: string, pattern: "^#[0-9A-F]{6}$"}
    MediaInputSource:
      type: object
      additionalProperties: false
      required: [id, type, connectionId]
      properties:
        id: {type: string, format: uuid}
        type: {type: string, enum: [MANUAL, CONNECTION]}
        connectionId: {type: [string, "null"], format: uuid}
    MediaInput:
      type: object
      additionalProperties: false
      required: [versionId, artifactId, role, order, color, sources]
      properties:
        versionId: {type: string, format: uuid}
        artifactId: {type: string, format: uuid}
        role: {$ref: "#/components/schemas/MediaInputRole"}
        order: {type: integer, minimum: 0}
        color: {type: string, pattern: "^#[0-9A-F]{6}$"}
        sources:
          type: array
          items: {$ref: "#/components/schemas/MediaInputSource"}
    PromptMediaMention:
      type: object
      additionalProperties: false
      required: [versionId, role]
      properties:
        versionId: {type: string, format: uuid}
        role: {$ref: "#/components/schemas/MediaInputRole"}
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId, retryable]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        traceId: {type: string}
        retryable: {type: boolean}
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
...
  responses:
    ValidationError:
      description: 请求参数无效
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Unauthenticated:
      description: 未登录或凭据不正确
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Forbidden:
      description: CSRF、初始化凭据或权限校验失败
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Conflict:
      description: 初始化或乐观版本冲突
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NotFound:
      description: 资源不存在或当前用户无权访问
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  headers:
    ContentLanguage:
      description: 协商后的界面语言；不改变用户内容、枚举或错误码。
      schema: {type: string, enum: [en, zh, ru, ja]}
    VaryLanguage:
      description: 包含 Accept-Language，避免不同语言的响应共用缓存。
      schema: {type: string}
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: 支持 en、zh、ru、ja 及其地区变体，按质量权重选择；缺失、不支持或无效时回退 zh。
      schema: {type: string}
      example: ru-RU, en;q=0.8
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    CanvasItemId:
      name: canvasItemId
      in: path
      required: true
      schema: {type: string, format: uuid}